// src/entities/payment/model/Payment.ts
export interface PaymentItem {
  id?: string;
  s_idx: number;
  ss_idx: number;
  b_idx: number;
  brand_code: string;
  amount: number;
  b_type: '지정지점' | '전지점';
  title: string;
  brand: string;
  branchType: string;
  duration: number;
  price: number;
  originalPrice?: number;
  sessions: number;
  name?: string;
  type: 'membership' | 'additional';
  // 추가관리 전용 필드
  mp_idx?: number;
  r_date?: string;
  r_memo?: string;
  branchName?: string;
}
// ...
export interface OrderResponse {
  resultCode: string;
  resultMessage: string;
  orderer: {
    csm_idx: string;
    name: string;
    hp: string;
    email: string;
  };
  orderSheet: {
    orderid: string;
    items: Array<{
      membership: {
        s_idx: string;
        s_name: string;
        s_time: string;
      };
      branch: {
        b_idx: string;
        b_name: string;
      };
      option: {
        ss_idx: string;
        ss_count: string;
      };
      origin_price: string;
      price: string;
      amount: number;
    }>;
  };
  orderSummary: {
    total_origin_price: number;
    total_price: number;
    total_count: number;
  };
  pg_info: {
    P_MID: string;
    P_OID: string;
    P_AMT: number;
    P_GOODS: string;
    P_UNAME: string;
    P_NEXT_URL: string;
    P_NOTI_URL: string;
    P_HPP_METHOD: string;
    P_RESERVED: string;
    P_TIMESTAMP: string;
    P_VBANK_DT?: string;
    P_VBANK_TM?: string;
  };
}

export interface BasePaymentParams {
  P_INI_PAYMENT: string;
  P_MID: string;
  P_OID: string;
  P_AMT: string;
  P_GOODS: string;
  P_UNAME: string;
  P_NEXT_URL: string;
  P_NOTI_URL: string;
  P_NOTI: string;
  P_CHARSET: string;
  P_HPP_METHOD: string;
  P_TIMESTAMP: string;
  P_RESERVED?: string;
  P_CARD_OPTION?: string;
  P_VBANK_DT?: string;
  P_VBANK_TM?: string;
  P_MOBILE?: string;
  P_APP_BASE?: string;
}